+++ packages/adminui-vue/CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.6.31]
+### Fixed
+ - duplication of store data in tabs in case store config is provided as object to `Form` with state as a function
+
 ### Added
- - export `SET` mutation from adminui-vue. It need when you use `computedVuex` in store module 
+ - `SET` mutation is exposed by adminui-vue. It need when you use `computedVuex` in store module 
+
 
 ## [1.6.30]
 ### Added